Reset to empty project to begin new build. Kept example under "app-example"
