+++ social-therapy-mobile/app-example/(tabs)/_layout.tsx
