+++ social-therapy-mobile/app-example/(tabs)/explore.tsx
