+++ social-therapy-mobile/app-example/(tabs)/index.tsx
